Reduce number of files needed to manage GUI
